Packets for Vending (minus wallet)

+++ src/main/java/gunn/modcurrency/mod/network/PacketSetLongToClient.java
@@ -20,12 +20,12 @@ import net.minecraftforge.items.IItemHandler;
  */
 public class PacketSetLongToClient implements IMessage {
     private BlockPos blockPos;
-    private long data;
+    private double data;
     private byte field;
 
     public PacketSetLongToClient(){}
 
-    public void setData(BlockPos pos, byte id, long data) {
+    public void setData(BlockPos pos, byte id, double data) {
         this.blockPos = pos;
         this.field = id;
         this.data = data;
@@ -36,7 +36,7 @@ public class PacketSetLongToClient implements IMessage {
     public void fromBytes(ByteBuf buf) {
         blockPos = new BlockPos(buf.readInt(), buf.readInt(), buf.readInt());
         field = buf.readByte();
-        data = buf.readLong();
+        data = buf.readDouble();
     }
 
     @Override
@@ -45,7 +45,7 @@ public class PacketSetLongToClient implements IMessage {
         buf.writeInt(blockPos.getY());
         buf.writeInt(blockPos.getZ());
         buf.writeByte(field);
-        buf.writeLong(data);
+        buf.writeDouble(data);
     }
 
     public static class Handler implements IMessageHandler<PacketSetLongToClient, IMessage> {
@@ -59,7 +59,7 @@ public class PacketSetLongToClient implements IMessage {
         private void handle(PacketSetLongToClient message, MessageContext ctx){
             final WorldClient worldClient = Minecraft.getMinecraft().world;
 
-            if(worldClient.getTileEntity(message.blockPos) instanceof TileVending)((TileVending) worldClient.getTileEntity(message.blockPos)).setLong(message.field, message.data);
+            if(worldClient.getTileEntity(message.blockPos) instanceof TileVending)((TileVending) worldClient.getTileEntity(message.blockPos)).setDouble(message.field, message.data);
         }
     }
